// src/components/ui/DatePicker/DatePicker.types.ts
export interface CalendarConfig {
  maxDate: string;
  minDate: string;
  defaultDate: string;
  startOfTheDay: boolean;
  endOfTheDay: boolean;
  startYear: number;
  endYear: number;
  startMonth: number;
  endMonth: number;
  useDateInput: boolean;
  useDatePicker: boolean;
  dateFormat: string;
  allowDateRange: boolean;
  setDefaultToday: boolean;
}